feat: modularize daily-entry screen components
Refactors the `daily-entry.tsx` file by extracting sub-components into their own dedicated files.

- Moves the `DynamicListSection` component to `components/daily-entry/DynamicListSection.tsx` for better reusability and separation of concerns.
- Moves the `RatingsSection` component to `components/daily-entry/RatingsSection.tsx`, making it a standalone component.
- Updates `app/daily-entry.tsx` to import and utilize the newly modularized components.
