+++ daily-planner/app/index.tsx
@@ -1,5 +1,5 @@
 // ============================================
-// app/index.tsx - REFACTORED
+// app/index.tsx
 // ============================================
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
